+++ Cargo.toml
@@ -4,7 +4,7 @@ description = "Write SQL queries in a simple and composable way"
 documentation = "https://docs.rs/sql_query_builder"
 repository = "https://github.com/belchior/sql_query_builder"
 authors = ["Belchior Oliveira <belchior@outlook.com>"]
-version = "1.0.3"
+version = "1.0.4"
 edition = "2021"
 rust-version = "1.58"
 license = "MIT"
