+++ tests/select_builder_api_spec.rs
@@ -12,8 +12,6 @@ fn all_clauses_concatenated_in_order() {
     .group_by("login")
     .having("active = true")
     .order_by("created_at desc")
-    .limit("1000")
-    .offset("50")
     .as_string();
 
   let expected_query = "\
@@ -24,9 +22,7 @@ fn all_clauses_concatenated_in_order() {
     WHERE user.login = $1 \
     GROUP BY login \
     HAVING active = true \
-    ORDER BY created_at desc \
-    LIMIT 1000 \
-    OFFSET 50\
+    ORDER BY created_at desc\
   ";
 
   assert_eq!(query, expected_query);
@@ -389,124 +385,6 @@ mod having_clause {
       .raw_after(sql::SelectClause::Having, "LIMIT 10")
       .as_string();
     let expected_query = "HAVING active = true LIMIT 10";
-
-    assert_eq!(query, expected_query);
-  }
-}
-
-mod limit_clause {
-  use super::*;
-  use pretty_assertions::assert_eq;
-
-  #[test]
-  fn method_limit_should_add_the_limit_clause() {
-    let query = sql::Select::new().limit("3").as_string();
-    let expected_query = "LIMIT 3";
-
-    assert_eq!(query, expected_query);
-  }
-
-  #[test]
-  fn method_limit_should_override_the_current_value() {
-    let query = sql::Select::new().limit("3").limit("4").as_string();
-    let expected_query = "LIMIT 4";
-
-    assert_eq!(query, expected_query);
-  }
-
-  #[test]
-  fn method_limit_should_trim_space_of_the_argument() {
-    let query = sql::Select::new().limit("  50  ").as_string();
-    let expected_query = "LIMIT 50";
-
-    assert_eq!(query, expected_query);
-  }
-
-  #[test]
-  fn clause_limit_should_be_after_order_by_clause() {
-    let query = sql::Select::new().order_by("created_at desc").limit("42").as_string();
-    let expected_query = "ORDER BY created_at desc LIMIT 42";
-
-    assert_eq!(query, expected_query);
-  }
-
-  #[test]
-  fn method_raw_before_should_add_raw_sql_before_limit_clause() {
-    let query = sql::Select::new()
-      .raw_before(sql::SelectClause::Limit, "group by id")
-      .limit("10")
-      .as_string();
-    let expected_query = "group by id LIMIT 10";
-
-    assert_eq!(query, expected_query);
-  }
-
-  #[test]
-  fn method_raw_after_should_add_raw_sql_after_limit_clause() {
-    let query = sql::Select::new()
-      .limit("10")
-      .raw_after(sql::SelectClause::Limit, "except select id, login")
-      .as_string();
-    let expected_query = "LIMIT 10 except select id, login";
-
-    assert_eq!(query, expected_query);
-  }
-}
-
-mod offset_clause {
-  use super::*;
-  use pretty_assertions::assert_eq;
-
-  #[test]
-  fn method_offset_should_add_the_offset_clause() {
-    let query = sql::Select::new().offset("100").as_string();
-    let expected_query = "OFFSET 100";
-
-    assert_eq!(query, expected_query);
-  }
-
-  #[test]
-  fn method_offset_should_override_the_current_value() {
-    let query = sql::Select::new().offset("100").offset("200").as_string();
-    let expected_query = "OFFSET 200";
-
-    assert_eq!(query, expected_query);
-  }
-
-  #[test]
-  fn method_offset_should_trim_space_of_the_argument() {
-    let query = sql::Select::new().offset("  2000  ").as_string();
-    let expected_query = "OFFSET 2000";
-
-    assert_eq!(query, expected_query);
-  }
-
-  #[test]
-  fn clause_offset_should_be_after_limit_clause() {
-    let query = sql::Select::new().limit("500").offset("100").as_string();
-    let expected_query = "LIMIT 500 OFFSET 100";
-
-    assert_eq!(query, expected_query);
-  }
-
-  #[test]
-  fn method_raw_before_should_add_raw_sql_before_offset_clause() {
-    let query = sql::Select::new()
-      .raw_before(sql::SelectClause::Limit, "limit 1000")
-      .offset("50")
-      .as_string();
-    let expected_query = "limit 1000 OFFSET 50";
-
-    assert_eq!(query, expected_query);
-  }
-
-  #[test]
-  fn method_raw_after_should_add_raw_sql_after_offset_clause() {
-    let query = sql::Select::new()
-      .offset("10")
-      .raw_after(sql::SelectClause::Offset, "/* the end */")
-      .as_string();
-    let expected_query = "OFFSET 10 /* the end */";
 
     assert_eq!(query, expected_query);
   }
